Valid Schedule Found
Currently cannot handle more than one processor - however it finishes one depth first search entirely without error - there is an issue when it tries to add = but will fix that soon.

However: returns a valid solution AND labels the node with start time and processor.

Needs to be refactored. but should be completed first.