Fixed arraylist length issue
getOptimal is now called each time a new solution is added to the array. which ensures that the arrayList doesn't ever get to big as the max length is now 2. (This may affect the time a lot though... as it will go through a lot more loops!!)

Scheduler now calls getGraph which labels the graph and returns it - no longer calls getOptimal as this is done continuously.