corrected SolutionSpaceIT
Created a new method in SolutionSpaceManager called getOptimalFinishTime() that returns an int to be used for testing - it returns the finish time of the optimal schedule.
Edited the testing class for this, added 5 new test cases and changed the existing ones to have more coverage. Values need to be approved for optimality and coverage of test cases should be further questioned.