Renamed SolutionSpaceManager to DFS Algorithm
Refactored DFSAlgorithm so that it extends Algorithm - removes duplicate code.
Deleted SolutionSpace.java as that code is no longer used (DFSAlgorithmIT) is a rewrite of it.

Changed calls to SolutionSpaceManager to DFSAlgorithm.

Removed  public abstract buildSolution() in Algorithm as DFS needs a param passed in while A* does not - so not consistent.