Singleton class renamed and DFS tests passing
Renamed the singleton class so that it can also be used for A* Algorithm.
DFSAlgorithmIT is updated so that it uses the new DFS Algorithm - testing is hard coded for one core at the moment though.
Added a few comments to explain the changes.
Further refactoring needed ie. deleting the old DFS algorithm class - but will discuss with group before proceeding