Changed return to System.exit
Changed return statements in scheduler to System.exit(1) when an error had occurred and changed to system.exit(0) when exiting after saving a file (no error).
Changed the reading in params error messages and ensured that exits main if 0 cores or processors are specified.
Added a default .dot to output files if a .dot was not already specified by the user.